refactor: move mutexkv to its own package
Moving mutexkv out of provider.go makes it easier to restructure
our code.
