+++ equinix/provider.go
@@ -22,7 +22,6 @@ import (
 )
 
 var (
-	metalMutexKV         = NewMutexKV()
 	DeviceNetworkTypes   = []string{"layer3", "hybrid", "layer2-individual", "layer2-bonded"}
 	DeviceNetworkTypesHB = []string{"layer3", "hybrid", "hybrid-bonded", "layer2-individual", "layer2-bonded"}
 	NetworkTypeList      = strings.Join(DeviceNetworkTypes, ", ")
